Change Transaction Id Type from Int to String
+ auto generate it by using the current time

+++ OwlioApp/app/src/main/java/com/example/owlio/ui/screen/form/EditTransactionFormScreen.kt
@@ -21,7 +21,7 @@ fun EditTransactionFormScreen(
 
     LaunchedEffect(key1 = transactionId) {
         // update the view model with the data for the current transaction
-        vm.fillFormFromId(transactionId.toInt())
+        vm.fillFormFromId(transactionId)
 
     }
